@@ -7,33 +7,40 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "compile": "tsc",
     "build-prod": "tsc",
-    "selenium": "node dist/benchmarkRunner.js",
-    "checkSelenium": "node dist/benchmarkRunner.js --check true",
+    "bench": "cross-env LANG=\"en_US.UTF-8\" node dist/benchmarkRunner.js",
+    "selenium": "cross-env LANG=\"en_US.UTF-8\" node dist/benchmarkRunner.js",
+    "checkSelenium": "cross-env LANG=\"en_US.UTF-8\" node dist/benchmarkRunner.js --check true",
     "results": "npm run interactive-results",
     "static-results": "echo \"no longer supported\" && exit 1",
     "interactive-results": "node dist/createResultJS.js && cd .. && cd webdriver-ts-results && npm run build-prod",
     "index": "node dist/createIndex.js",
     "check": "node dist/runCheck.js",
-    "nonKeyed": "node dist/nonKeyed.js"
+    "update": "node dist/updateFrameworks.js",
+    "nonKeyed": "cross-env LANG=\"en_US.UTF-8\" node dist/isKeyed.js",
+    "isKeyed": "cross-env LANG=\"en_US.UTF-8\" node dist/isKeyed.js"
   },
   "author": "",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/lodash": "4.14.104",
-    "@types/node": "9.4.6",
-    "@types/ramda": "^0.25.19",
+    "@types/lodash": "4.14.108",
+    "@types/node": "10.0.8",
+    "@types/ramda": "^0.25.28",
     "@types/selenium-webdriver": "3.0.8",
+    "@types/semver": "^5.5.0",
     "@types/yargs": "11.0.0",
-    "typescript": "2.7.2"
+    "typescript": "2.8.3"
   },
   "dependencies": {
-    "chromedriver": "2.36.0",
+    "cross-env": "5.1.1",
+    "chromedriver": "2.38.3",
     "dot": "1.1.2",
     "jstat": "1.7.1",
-    "lighthouse": "2.9.1",
-    "lodash": "4.17.5",
+    "lighthouse": "2.9.4",
+    "lodash": "4.17.10",
     "ramda": "^0.25.0",
     "selenium-webdriver": "4.0.0-alpha.1",
-    "yargs": "11.0.0"
+    "yargs": "11.0.0",
+    "npm-check-updates": "^2.14.2",
+    "semver": "^5.5.0"
   }
 }
