@@ -2,30 +2,69 @@ var _ = require('lodash');
 var exec = require('child_process').execSync;
 var fs = require('fs');
 var path = require('path');
+var yargs = require('yargs');
 
-var excludedDirectories = ['css', 'images', 'dist','node_modules','webdriver-java'];
+let args = yargs(process.argv)
+    .usage("npm run build [-- [--check] [--skipIrrelevant] [--restartWith]]")
+    .help('help')
+    .boolean('check')
+    .boolean('skipIrrelevant')
+    .string('restartWith')
+    .argv;
 
-// set the following variable to resume building with a framework and skip all
-// other frameworks that would be buils before
-var restartWithFramework = '';
-var build = !restartWithFramework ? true : false;
+var referenceBranch = "origin/master";
 
-_.each(fs.readdirSync('.'), function(name) {
-	if(!name.startsWith(".") && fs.statSync(name).isDirectory() && excludedDirectories.indexOf(name)==-1 && fs.existsSync(path.join(name, "package.json"))) {
-		if (!build && name.startsWith(restartWithFramework)) build = true;
-		if (build) {
-			console.log("*** Executing npm install in "+name);
+var restartWithFramework = args.restartWith || '';
+
+var core = ["webdriver-ts", "webdriver-ts-results"].map(f => ["", f]);
+
+var frameworks = [].concat(
+  fs.readdirSync('./frameworks/keyed').map(f => ['frameworks/keyed/', f]),
+  fs.readdirSync('./frameworks/non-keyed').map(f => ['frameworks/non-keyed/', f]));
+
+var notRestarter = ([_, name]) => !name.startsWith(restartWithFramework || undefined);
+var [skippable, buildable] = !restartWithFramework
+    ? [[],
+       frameworks]
+    : [_.takeWhile(frameworks, notRestarter),
+       _.dropWhile(frameworks, notRestarter)];
+
+var relevant = args.skipIrrelevant && !_.some(core, isDifferent)
+    ? _.filter(buildable, isDifferent)
+    : buildable;
+
+_.each(skippable, ([dir,name]) => console.log("*** Skipping " + dir + name));
+
+_.each([].concat(relevant, core), function([dir,name]) {
+        let fullname = dir + name;
+	if(fs.statSync(fullname).isDirectory() && fs.existsSync(path.join(fullname, "package.json"))) {
+          console.log("*** Executing npm install in "+fullname);
             exec('npm install', {
-				cwd: name,
+				cwd: fullname,
 				stdio: 'inherit'
 			});
-			console.log("*** Executing npm run build-prod in "+name);
+			console.log("*** Executing npm run build-prod in "+fullname);
 			exec('npm run build-prod', {
-				cwd: name,
+				cwd: fullname,
 				stdio: 'inherit'
 			});
-		} else {
-			console.log("*** Skipping "+name);
-		}
 	}
 });
+
+var testable = args.check ? relevant : [];
+_.each(testable, function([dir,name]) {
+        let fullname = dir + name;
+	if(fs.statSync(fullname).isDirectory() && fs.existsSync(path.join(fullname, "package.json"))) {
+            console.log("*** Executing npm run selenium for "+fullname);
+            exec('npm run selenium -- --count 1 --fork false --framework ' + name, {
+				cwd: "webdriver-ts",
+				stdio: 'inherit'
+			});
+	}
+});
+
+function isDifferent([dir,name]) {
+  try { exec('git diff --quiet ' + referenceBranch + ' -- ' + dir + name); }
+  catch(e) { return true; }
+  return false;
+};
