@@ -3,38 +3,31 @@ var exec = require('child_process').execSync;
 var fs = require('fs');
 var commandExists = require('command-exists');
 const path = require('path');
-
-var installCommand = 'npm install';
-
-var excludedDirectories = ['css', 'dist','node_modules','webdriver-java'];
+const rimraf = require('rimraf');
 
 function rmIfExists(base, name) {
 	let dir = path.join(base, name);
 	if(fs.existsSync(dir)) {
-		let cleanCommand = 'rm -r '+name;
-		console.log("Executing "+cleanCommand+" in "+base);
-		exec(cleanCommand, {
-			cwd: base,
-			stdio: 'inherit'
-		});
+		console.log("Clean ",dir);
+        rimraf.sync(dir);
 	}
 }
 
-_.each(fs.readdirSync('.'), function(name) {
-	if(fs.statSync(name).isDirectory() && name[0] !== '.' && excludedDirectories.indexOf(name)==-1) {
-		if(fs.existsSync(name+"/node_modules")) {
-			let cleanCommand = 'rm -r node_modules';
-			console.log("Executing "+cleanCommand+" in "+name);
-			exec(cleanCommand, {
-				cwd: name,
-				stdio: 'inherit'
-			});
+for (let keyedType of ['keyed', 'non-keyed']) {
+    let dir = path.resolve('frameworks', keyedType);
+    let directories = fs.readdirSync(dir);
+
+    for (let name of directories) {
+        let fd = path.resolve(dir, name);
+        console.log('cleaning ', fd);
+		if(fs.existsSync(fd+"/node_modules")) {
+			rimraf.sync(fd+"/node_modules");
 		}
-		rmIfExists(name, "package-lock.json");
-		rmIfExists(name, "yarn.lock");
-		rmIfExists(name, "dist");
-		rmIfExists(name, "elm-stuff");
-		rmIfExists(name, "bower_components");
-		rmIfExists(name, "node_modules");
+		rmIfExists(fd, "package-lock.json");
+		rmIfExists(fd, "yarn.lock");
+		rmIfExists(fd, "dist");
+		rmIfExists(fd, "elm-stuff");
+		rmIfExists(fd, "bower_components");
+		rmIfExists(fd, "node_modules");
 	}
-});
+}
