@@ -2,7 +2,7 @@ var _ = require('lodash');
 var exec = require('child_process').execSync;
 var fs = require('fs-extra');
 var path = require('path');
-var yargs = require ('yargs'); 
+var yargs = require ('yargs');
 
 let args = yargs(process.argv)
 .usage("$0 [--bootstrap --minimal]")
@@ -39,7 +39,7 @@ fs.copySync("index.html", "dist"+path.sep+"index.html");
 fs.copySync("css", "dist"+path.sep+"css");
 
 var excludes = ["node_modules","elm-stuff","project",".DS_Store"]
-var excludedDirectories = ['css', 'dist','node_modules','webdriver-java', 'webdriver-ts'];
+var excludedDirectories = ['css', 'dist','node_modules','webdriver-ts'];
 
 // http://stackoverflow.com/questions/13786160/copy-folder-recursively-in-node-js
 function copyFileSync( source, target ) {
@@ -63,11 +63,11 @@ function include(name) {
 				}
 		}
 		if (excludes.every(ex => name.indexOf(ex)==-1)) {
-			// console.log("<- filter", name);					
+			// console.log("<- filter", name);
 			return true;
 		} else {
 			return false;
-		}	
+		}
 }
 
 function copyFolderRecursiveSync( source, target ) {
