@@ -1,6 +1,10 @@
 import * as fs from 'fs';
-import {JSONResult, config, frameworks} from './common'
+import {JSONResult, config, initializeFrameworks} from './common'
 import {BenchmarkType, Benchmark, benchmarks} from './benchmarks'
+
+let frameworks = initializeFrameworks();
+
+frameworks.sort( (a,b) => a.fullNameWithKeyedAndVersion.localeCompare(b.fullNameWithKeyedAndVersion));
 
 const dots = require('dot').process({
 	path: './'
